Move doc parsing related deps out of the core
The reason we don't want these in the core module is because
not every application needs the doc parsing functionality
(which adversely affects the application size).
For example, a simple AiService that uses OpenAI  has at least 30MB
added to binary size by just including the POI and BouncyCastle dependencies.
