+++ easy-rag/deployment/pom.xml
@@ -12,7 +12,7 @@
 
         <dependency>
             <groupId>io.quarkiverse.langchain4j</groupId>
-            <artifactId>quarkus-langchain4j-core-deployment</artifactId>
+            <artifactId>quarkus-langchain4j-parsers-base-deployment</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
