+++ parsers-base/deployment/src/main/java/io/quarkiverse/langchain4j/deployment/parsers/base/DocumentSupportProcessor.java
@@ -1,18 +1,11 @@
-package io.quarkiverse.langchain4j.deployment;
+package io.quarkiverse.langchain4j.deployment.parsers.base;
 
-import static io.quarkiverse.langchain4j.deployment.JarResourceUtil.determineJarLocation;
-import static io.quarkiverse.langchain4j.deployment.JarResourceUtil.matchingJarEntries;
-
-import java.nio.file.Path;
 import java.util.List;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import java.util.zip.ZipEntry;
 
 import org.apache.commons.compress.compressors.CompressorStreamFactory;
 import org.jboss.logmanager.Level;
 
-import dev.langchain4j.data.document.splitter.DocumentBySentenceSplitter;
 import dev.langchain4j.rag.query.transformer.CompressingQueryTransformer;
 import io.quarkiverse.langchain4j.deployment.items.InProcessEmbeddingBuildItem;
 import io.quarkus.bootstrap.classloading.QuarkusClassLoader;
@@ -103,21 +96,6 @@ public class DocumentSupportProcessor {
                     .fields(true)
                     .methods(true)
                     .build());
-        }
-    }
-
-    @BuildStep
-    void openNLPResources(
-            BuildProducer<NativeImageResourceBuildItem> producer) {
-        registerCustomOpenNLPResources(producer);
-    }
-
-    private void registerCustomOpenNLPResources(BuildProducer<NativeImageResourceBuildItem> resourcesProducer) {
-        Path langChain4jJar = determineJarLocation(DocumentBySentenceSplitter.class);
-        List<String> names = matchingJarEntries(langChain4jJar, e -> e.getName().endsWith(".bin")).stream().map(
-                ZipEntry::getName).collect(Collectors.toList());
-        if (!names.isEmpty()) {
-            resourcesProducer.produce(new NativeImageResourceBuildItem(names));
         }
     }
