+++ parsers-base/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,0 +1,10 @@
+name: Quarkus LangChain4j Parsers Base
+artifact: ${project.groupId}:${project.artifactId}:${project.version}
+description: Provides the base module for making LangChain4j parsers work in Quarkus LangChain4j
+metadata:
+   keywords:
+    - ai
+    - langchain4j
+   categories:
+     - "miscellaneous"
+   status: "experimental"
