+++ samples/csv-chatbot/src/main/resources/application.properties
@@ -1,6 +1,8 @@
 quarkus.langchain4j.openai.timeout=60s
 csv.file=src/main/resources/data/movies.csv
 
+quarkus.hibernate-orm.database.generation=drop-and-create
+
 # if you want to log the requests and responses that go to OpenAI:
 #quarkus.langchain4j.openai.log-requests=true
 #quarkus.langchain4j.openai.log-responses=true
